refactor: helm pipeline and helm chart controllers
