+++ internal/controller/helm_chart/helmchart_controller_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package controller
+package helm_chart
 
 // import (
 // 	"context"
