+++ internal/controller/helm_chart/templates/helm-install-job.yml.tpl
