+++ internal/controller/helm_chart/templates/helm-uninstall-job.yml.tpl
